requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
openai==1.12.0
pysnow==0.7.17
langchain==0.1.9
langchain-core==0.1.27
langchain-openai==0.0.8
langchain-community
pydantic==2.6.1
python-multipart==0.0.6
aiohttp==3.9.3